can_report/requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
dotenv==0.9.9
et_xmlfile==2.0.0
idna==3.10
numpy==2.2.5
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
smartsheet-python-sdk==3.0.5
tzdata==2025.2
urllib3==2.4.0
